Create btapneaservice.java
changed boolean waitforack to volatile - still no difference
read online that The Java memory model guarantees that each thread entering a synchronized block 
of code sees the effects of all previous modifications that were guarded by the same lock - still no diff 
okay so, classification which is done in the intent thread takes some time and after that the message is sent to 
the arduino, in the meantime read has already been called in connect thread  
can do - thread sleep for till next sec
now it is read at the right time : after being set, but still not working
created local var which is set to waitforack
works! but failure still possible if threads interleave in a certain manner
failed! that sort of interleaving did happen see log4.txt